refactor: isolate the disabled lane queue
